# gateway/api/gateway/v1/gateway_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v4.25.3
// source: gateway/v1/gateway.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v11 "proto_definitions/product/v1"
	v12 "proto_definitions/seckill/v1"
	v1 "proto_definitions/user/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGatewayServiceGetProductInfo = "/api.v1.GatewayService/GetProductInfo"
const OperationGatewayServiceLogin = "/api.v1.GatewayService/Login"
const OperationGatewayServiceSeckill = "/api.v1.GatewayService/Seckill"

type GatewayServiceHTTPServer interface {
	GetProductInfo(context.Context, *v11.QueryRequest) (*v11.ProductInfoResponse, error)
	Login(context.Context, *v1.LoginRequest) (*v1.LoginResponse, error)
	Seckill(context.Context, *v12.SeckillRequest) (*v12.SeckillResponse, error)
}

func RegisterGatewayServiceHTTPServer(s *http.Server, srv GatewayServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/user/login", _GatewayService_Login0_HTTP_Handler(srv))
	r.GET("/product/info/{id}", _GatewayService_GetProductInfo0_HTTP_Handler(srv))
	r.POST("/seckill", _GatewayService_Seckill0_HTTP_Handler(srv))
}

func _GatewayService_Login0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*v1.LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.LoginResponse)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_GetProductInfo0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.QueryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceGetProductInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetProductInfo(ctx, req.(*v11.QueryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ProductInfoResponse)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_Seckill0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v12.SeckillRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceSeckill)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Seckill(ctx, req.(*v12.SeckillRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v12.SeckillResponse)
		return ctx.Result(200, reply)
	}
}

type GatewayServiceHTTPClient interface {
	GetProductInfo(ctx context.Context, req *v11.QueryRequest, opts ...http.CallOption) (rsp *v11.ProductInfoResponse, err error)
	Login(ctx context.Context, req *v1.LoginRequest, opts ...http.CallOption) (rsp *v1.LoginResponse, err error)
	Seckill(ctx context.Context, req *v12.SeckillRequest, opts ...http.CallOption) (rsp *v12.SeckillResponse, err error)
}

type GatewayServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewGatewayServiceHTTPClient(client *http.Client) GatewayServiceHTTPClient {
	return &GatewayServiceHTTPClientImpl{client}
}

func (c *GatewayServiceHTTPClientImpl) GetProductInfo(ctx context.Context, in *v11.QueryRequest, opts ...http.CallOption) (*v11.ProductInfoResponse, error) {
	var out v11.ProductInfoResponse
	pattern := "/product/info/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayServiceGetProductInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *GatewayServiceHTTPClientImpl) Login(ctx context.Context, in *v1.LoginRequest, opts ...http.CallOption) (*v1.LoginResponse, error) {
	var out v1.LoginResponse
	pattern := "/user/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGatewayServiceLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *GatewayServiceHTTPClientImpl) Seckill(ctx context.Context, in *v12.SeckillRequest, opts ...http.CallOption) (*v12.SeckillResponse, error) {
	var out v12.SeckillResponse
	pattern := "/seckill"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGatewayServiceSeckill))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}