# gateway/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: GatewayService API
    version: 0.0.1
paths:
    /product/info/{id}:
        get:
            tags:
                - GatewayService
            operationId: GatewayService_GetProductInfo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.ProductInfoResponse'
    /seckill:
        post:
            tags:
                - GatewayService
            operationId: GatewayService_Seckill
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/seckill.v1.SeckillRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/seckill.v1.SeckillResponse'
    /user/login:
        post:
            tags:
                - GatewayService
            operationId: GatewayService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LoginResponse'
components:
    schemas:
        product.v1.ProductInfoResponse:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                describe:
                    type: string
                price:
                    type: string
                stock:
                    type: integer
                    format: int32
        seckill.v1.SeckillRequest:
            type: object
            properties:
                userID:
                    type: string
                productID:
                    type: string
        seckill.v1.SeckillResponse:
            type: object
            properties:
                orderID:
                    type: string
                userID:
                    type: string
                productID:
                    type: string
                price:
                    type: string
        user.v1.LoginRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        user.v1.LoginResponse:
            type: object
            properties:
                token:
                    type: string
tags:
    - name: GatewayService